refactor (Customer) : moving files to package from context
