+++ grails-app/controllers/com/mini/asaas/payer/PayerController.groovy
@@ -2,7 +2,7 @@ package com.mini.asaas.payer
 
 import com.mini.asaas.payer.Payer
 import com.mini.asaas.payer.PayerService
-import com.mini.asaas.Customer
+import com.mini.asaas.customer.Customer
 import com.mini.asaas.dto.payer.PayerDTO
 import grails.validation.ValidationException
 import grails.compiler.GrailsCompileStatic
