+++ grails-app/domain/com/mini/asaas/payer/Payer.groovy
@@ -1,7 +1,7 @@
 package com.mini.asaas.payer
 
 import com.mini.asaas.base.PersonEntity
-import com.mini.asaas.Customer
+import com.mini.asaas.customer.Customer
 import grails.compiler.GrailsCompileStatic
 
 @GrailsCompileStatic
