+++ grails-app/services/com/mini/asaas/customer/CustomerService.groovy
@@ -1,6 +1,6 @@
-package com.mini.asaas
+package com.mini.asaas.customer
 
-import com.mini.asaas.Customer
+import com.mini.asaas.customer.Customer
 import com.mini.asaas.Address
 import com.mini.asaas.AddressService
 import com.mini.asaas.dto.customer.CustomerDTO
