+++ grails-app/services/com/mini/asaas/payment/PaymentService.groovy
@@ -1,7 +1,7 @@
 package com.mini.asaas.payment
 
 import com.mini.asaas.payment.Payment
-import com.mini.asaas.Customer
+import com.mini.asaas.customer.Customer
 import com.mini.asaas.payer.Payer
 import com.mini.asaas.dto.payment.CreatePaymentDTO
 import com.mini.asaas.dto.payment.UpdatePaymentDTO
